✨ Use the DevQuery API to access the Device Container properties.
This commit adds P/Invoke for the entire API surface of DevQuery, which does not seem to be documented outside of header defintions?
DevQuery API turns out to be way nicer (it is what seems to be used as the core of Windows.Devices.Enumeration) than the old CM and SetupApi APIs. It might be possible to entirely get rid of these APIs in the code, but ot sure if there are any downsides to getting rid of them yet. Except for the time wasted understanding/exposing/using them, of course 😞
Due to its async nature, DevQuery might even be able to entirely replace the tedious Device Notification stuff that is to be used in the Windows Service.