🎉 First draft of exposing DevQuery APIs in an usable way.
This turned out to be way more complex than it should have been, because these APIs being only half documented.
i.e. The prototypes are defined, but there is no official documentation outside of the WinRT wrapper Windows.Devices.Enumeration which is too high-level (and bloated) for us.
Anyway, with this first step, it should be possible to provide a nicer access to this API surface.