✨ Add the overlay UI.
This introduces many necessary structural changes, and adds the overlay UI taht will be used to display various useful on-screen notifications.
While this is plugged and somewhat working, the overlay Window placement is not done properly yet.
Many other changes are also required (variable font, proper queuing/overriding of overlays) but not critical for the first version of the code.