✨ Add overlay support for caps lock and others.
No overlay show for num lock and scroll lock for now due to the lack of a suitable glyph.
This also only works with logitech devices for now, as they directly report these notifications in the protocol. Doing it for all keyboards regardless of the brand should be possible but with a totally different approach.
We may still want to display generic (non-device specific) overlays though, but it will comme later in any case.