✨ Prevent exceptions in UI watcher.
If the call to GetExtendedDeviceInformationAsync fails, it could lead to the battery level to stay in the dictionary. Reason for the call to fail is still weird, but it is supposed to happen at times if a device disconnects too fast. (From waht I saw it always failed after the device was added, when the device should have been available. But it was not. But it went up right afterwards… So maybe some kind of device connection debounce is needed?)