♻️ Refactoring for persisting configuration.
This needs more work but this was taking a long time and already a lot of stuff has been touched, so this is a good stopping point.
Device registry likely needs to be made full async next. And compatible device instance indexes need to be handled, which turns out to be a not-so-nice nice problem to solve. The IndexedConfigurationKey might or might not go away in the process.
Basically, the idea is that we never need to index the comaptible devices when they have a serial number (which we can hopefully retrieve quite often), but we still need the comaptibility ID to allow the user to manually reassign a device configuration later on. (e.g. when replacing a device by a new one)