✨ Start implementing async ioctl.
This stuff seems complicated to do right, but it will be very useful.
It will likely only work on .NET 8.0+, as previous versions would require to put in more work due to the missing features. (Mostly convenience stuff is missing here)
First step would probably be to design a "DeviceStream" based on FileStream and exposing the ioctl stuff, then use that to provide async services in HidFullDuplexStream. Providing the feature separately from a stream can work but with some overhead, as objects will need to be allocated everytime.