✨ Retrieve HID preparsed data asynchronously.
Especially, we get back full access to the preparsed data bytes, as we know the length and allocate the array ourselves. (While the HID++, unlike the RawInput API, would allocate an opaque memory block to hold the data)
Also add an HidDeviceStream class to hold some of the logic, as everything is based on ioctl anyway.