✨ Start wiring up the programming service.
Not doing anything useful in itself yet, but one of the many steps towards the feature:
Programming service is wired in order to consume the messages that are emitted, so as to avoid a memory leak. But the messages are ignored.
Next step would be to migrate the overlay logic into the service in the form of hardcoded programming for now.
Target implementation would probably generate a default program during the first run, so that minimal programmed features are readily available, but the user would be able to disable or customize everything.