✨ Implement basic mouse DPI notifications.
These are hardcoded for now, of course.
When available, the current level and number of DPI levels should be passed in the notification so that the level can be displayed using the level bar.
The overlay notification should probably be augmented with an actual value field in order to pass both the level and real value.