♻️ Refactor the discovery engine, 90% done.
All is mostly done to the plan that was made earlier.
Logs are mostly missing and need to be reimplemented, some small but important features still need to be implemented, and some critical bugs still need to be fixed.
However, this version starts mostly properly. It NEEDS to be improved, but it is in a workable state.