✨ Add/move standard effects with speed setting.
NB: Effect stuff need to be refactored a bit in the future. What I had in mind a few months ago was auto-converting simpler standard effects into their more complex (device-specific) form once they are applied. There might other options.
Still need to keep the standard effect stuff anyway, because it will definitely be useful for the programming stuff, to be able to assign a simple effect to any supporting device without knowing the specifics of that device.