♻️ Rename effects so that `Color` is only used for effects where a color is actually configurable.
e.g. While color cycle is generally understandable, it starts being a problem when we have `ColorWave`, that could actually be a wave one one color, with only brightness changes.
I don't particularly like the use of the word spectrum, but it is more generic than rainbow, and easily understandable, so here it is 🤷‍♂️