♻️ Start implementing color arrays for color effects.
Having a property for each color was unworkable and as consequence, the UI for that was simultaneously not good and quite a bit slow. (Each colorpicker currently being instanciated separatly, successive instanciations were becoming slower and slower)
The UI side still needs to be adapted, as for now effects using color arrays will throw an exception, but the backend support is here.