✨ Small tweaks to AsyncLock.
There was possibly a (never encountered) bug when the wait was cancelled. This change should address that.
A further change to replace TaskCompletionSource with IValueTaskSource seems doable to reduce allocations to the minimum in case of contention too. Value not certain as most uses are (hopefully) not contended, but it could still be useful.