♻️ Split the metadata service in each category, also according to waht was done in the metadata compiler.
The idea is that not all packages will provide all kinds of metadata, so we can certainly avoid iterating on too many archives.
There is also still the possibility of consolidating archives  during the initial run phase, but that would not go that well with the ability to dynamically load or unload plugins. Also not possible if we don't know the keys of everything (which we lose during the creation process, as we don't really need it)
Maybe a better alternative would be to keep track of actually requested data and use it to build a consolidated "live" archive that will contain data that is really used and needed on the machine. That would be moderately cheap to do, as we'd only need to keep track of the requested keys somewhere.
Anyway, this is for later.