🐛 HID++: Address possible conflicts with G HUB.
The code reading device responses for HID++ 1.0 could incorrectly match responses from other softwares.
Some queried registers include extra parameters which were not validated to be the same between request and response. This lead to problems when reading the response for the request of another software.
This is less likely to happen for HID++ 2.0, as the requests include an app i. However, all receivers are still based on HID++ 1.0.