♻️ Factorize code in monitor service.
This will require slightly more allocations because of the delegates used, but code duplication in much reduced.
An alternative would be to have a generic "IMonitorVcpFeature" interface from which all others are derived, but it would mean that on driver-side, all features would need to have their own instances (and class if it cannot be shared).
Might chaneg this in the future.