✨ Add the monitor control proxy. (1/N)
This feature is important for all setups where we won't have GPU drivers to access monitor I2C channels.
It will allow to run basic monitor DDC commands using the Windows API from the helper program started in interactive mode (Exo.Overlay).
Without that, it is impossible to use monitor DDC features, as the Windows API only works for programs started in a graphical user session. (Which is never the case for services)
For now, the basic code is mostly functional, but there are a few problems that prevent the feature from working:
- Sadly, the model I used for proxying requests is not good enough for GRPC so it won't work. (Can't have parameters together with streaming, and the CallOptions/ServerCallContext stuff looks plain terrible, as most of the GRPC concepts do.)
- The monitor control proxy is not yet actually used. We don't want to override GPU drivers if not necessary, but we still want to be able to provide support for monitors at some point.
- Silent reconnection to the proxied monitor on the service-side is not implemented. (Meaning that a disconnect will permanently affect a monitor device)