✨ Add the monitor control proxy. (3/N)
This adds the reconnection logic (untested yet) and allows pluging the fallback I2C bus provider into any display adapter driver that requires it.
NB: May want to move the reconnect logic in another part of the code. It is currently implemented in the fallback feature, within the I2CBus wrapper, while retrospectively, a better place to have it would be in a wrapper around the IMonitorControlService implementation or  possibly in the implementation itself.