✨ Persist loaded assemblies.
This was a shortseeing of the initial implementation of metada that sould have been corrected earlier, but it was not strictly needed.
Basically, we need to be able to access the metadata related to a device that has been seen in the past but is not connected at the moment. However, the assembly containing the driver for that device is likely to not be loaded (for memory saving purposes).
It will especially be a problem to order lighting zones in the UI.