🛠️ IT5702: Small tweak to the unified lighting mode apply.
It shouldn't change anything visually, but it is more in line with what is done by the official tool.
Still can't get the save settings command to work while I'm sure I witnessed it working in the past, but from my understanding it should require a single (simple) effect to be set on all zones, hence the tweaks to that part.