✨ Add ReceiveInputReportAsync to HID streams.
As it turns out, some devices do not push messages to the host. In that case, GET_REPORT must be used in the same way as for feature reports…
Having this function exposed here will serve that purpose, and maybe some niche adjacent use cases.
Il all other cases, we're perfectly fine with reading from the stream, and should continue doing so.