✨ Mouse polling frequency updates in the UI.
With this, most absic mouse features are supported.
The next step would be to implement support for advanced mouse profiles, which is needed but is of a higher complexity level. (Current features may also be nested within mouse profiles)
Supporting complex profiles will be needed to properly support logitech devices anyway.