✨ (Finally) Implement custom menu edition synchronization with the service.
UI for editiong a menu was already available but it didn't work on additional data.
Now, the UI allows for editing the actually used menu an push the changes to the backend.
Ability to detect change state and to revert changes from the UI is not yet implemented, but that doesn't prevent making changes.