✨ HID++: Start implementing color LED effects feature.
Mostly type declarations for now. Would need to have a device to properly try it. (Turns out G Pro X Wireless exposes the feature as engineering and hidden, so it is impossible to know if it actually that or a conflicting ID. Requests throw an error indicating logitech internal feature.)