✨ HID++: Fix profile name encoding: Use UTF-16.
While this is not documented, the default profile seemed to use UTF-16 encoding to encode "PROFILE_NAME_DEFAULT", and from testing, the OnboardMemoryManager software will only interpret the names as UTF-16.
So, while the field seems to be freeform, its format has an importance (Also maybe some devices use it?).