♻️ Allow matching multiple devices to the same main device name.
This will address the scenarios of switching between different setups of identical hardware (e.g. docking stations) or anything similar.
It will be strictly impossible for any two of those devices to be connected at the same time, so it will not be a problem.
Two of those devices can still be connected at the same time if they use a different port. Then, the main device name remapping logic will kick in as expected.