🐛 Fix the Razer HyperSpeed dongle.
Basically, stuff would only work if the wireless device was online, which is often the case, but not guaranteed. In which case, the dongle initialization failed for a few reasons.

Reason 1: The handshake stuff that was (re-)enabled in the code is actually not a handshake and must be answered by the device. (Uninvestigated keyboard layout stuff)
Reason 2: The device flags applied to the dongle were still that of the device, but since the last refactorings, these flags lead to some requests being sent to the dongle. (Which would be fine if the device was online, but wrong nonetheless)