🔩 Add automatic retries with delay for some DDC I2C reads.
Especially when the monitor gets connected, there can be some I2C usage conflicts, which would lead to receiving a response for another request. (Because when possible, we directly access the I2C instead of relying on Windows user APIs)
This can also happen randomly, or more likely if playing video because of HDCP traffic.