🐛 Force package version of Grpc.AspNetCore.Server.
This newer version (which is already used for the GRPC client) lowers the importance of GRPC cancellations so the log will be less polluted.
Some spurious error are still possible (trying to write on a completed channel in async enumerators), so the release log filter will be kept at `Fatal` for now.