✨ Cooling: Implement hardware curves in the backend service.
There is a small caveat at the moment, in that the data type of a sensor is considered to be identical for both sensor readings and cooling curve input.
In practice, it is possible for a sensor to have more precision than what the hardware cooling curve allows.
Since we hard-cast out cooling curves upon use in the backend, that shouldn't be too much of a problem, but it would be better if the information was (properly) transferred to the UI.
This is to be done later, as it is not critical for minimum operation.