🎨 Finally fix the ticks layout in the control curve editor.
This was far from critical but it was utterly ugly.
There might be better solutions than the one currently impelmented, but gives the expected results, so that would be good enough.

Had to resort to Binding with TemplatedParent instead of TemplateBinding (bug?) at many places though. But again, it works…