✨ Add a cache to the filename to BitmapImage converter.
I noticed that the filename property would be read multiple times in succession on the same instance. This should at least make this scenario less heavy on resources. This can be improved later on to have some kind of minimum retention timer for images, in order to mitigate the potentially too efficient GC. (If needed)