♻️ Try improving the transformation process in case of GIF.
Sadly, it still seems that some generated GIFs break the Kraken Z while they do play fine in apps.
I'm now suspecting that the encoding of frames by ImageSharp being far from optimal is the problem. Valid is not enough (if it is)
NZXT CAM probably uses some kind of lossy LZW encoding to reduce frame data.
Symptom on the device is that some frames take longer to render, and we can observe the animation "resetting". Which I'm assuming is caused by two operations racing against eachother. One being the frame rendering supposedly being done in real time from the encoded frame data. The other being the animation timer.
When the animation timer gets sufficiently ahead of the frame rendering, the GIF gets frozen, with the display possibly being corrupted. (Depends on the gif)
This does not seem to brick the FW though. Displaying another image afterwards is still possible.