✨ Implement in-memory caching of image transformations.
This is just a mapping between transformation parameters to the target image. The image data is not explicitly kept in memory by the application. (That's still managed by the OS)
However, we can avoid spending CPU transforming the image everytime, which means that swapping images on Kraken Z devices can now be ultra fast 😊