✨ Persist image transformation metadata 🎉
This is the last bit that was required to avoid recomputing images when the service is restarted.
An image that has already been transformed according to the requested transformation for a device will now always be recovered from the disk if it exists.
Code is imperfect (async void) but robust enough to do exactly what is required.