🎨 Add the ability to display notifications in the settings UI.
This will be useful to provide feedback in case of problems surfaced by the backend.
Specific handling will be added for each action in the UI. For now, I added it only to the image upload feature.
Obviously, the backend doesn't return any useful message currently, but we can always fix that later.