🛠️ Tweak the code to help freeing the memory used for processing images.
This is still not perfect, but it seems to be recovering altmost all of the memory from .NET POV at least.
Windows still sees a higher memory usage, which I'm assuming is caused by fragmentation. (The MemoryStream used to write the image is clearly not a good solution TBH. Getting rid of it would likely help)
This uses SimpleGcMemoryAllocator for now in order to avoid requesting a lot of shared arrays from .NET, but hopefully, this tweak will go away in the future.