🐛 Woops^2 😅
The changed bit is never called in practice but it could be.
And, well, there is still one way to mess up things and avoid freeing the buffer by misusing everything, but this should never happen if the APIs are used correctly.
Could be addressed by replacing the OR with a subtraction, and also do the dispose check on increment. Is it necessary though?