🛠️ Remove Kestrel port binding.
Ultimately, the goal would be to get rid of ASP.NET Core in its entirety, as it is in fact, not needed.
However, GRPC needs to be replaced by something else (hopefully simpler). While that is definitely something I want to do, it will be done when it will be done.