🎨 Reset the UI state of lamps when the device becomes offline.
This basically resyncs the internal actual device state with the requested state so that when the device comes back on, the actual device state can be displayed.
There is still the need for a similar mechanism overall so that the displayed state is force updated at some point when the device is still online (especially because color temp can easily desync), but it will come later on.