✨ Add power notifications so that lighting effects can be restored on wake from sleep.
This ended up far bigger than I expected.
Nothing that complicated in the end, but quite a log of small chanegs in addition to the main chunk.
Also includes a few other tweaks in the cusomized ServiceBase. I'll probably tune this one even more later on, but that was outside of the scope of this.
There is working code for restoring lighting settings (to test the notifications at least) but the driver contract should ideally be amended in order to be able to force devices to resend updates. (Otherwise, they could choose to trust their internal state)
Ideally, we could even avoid re-setting the effect on the device, as the current contract forces all drivers to keep their current state cached anyway. Also probably let the driver tell us if it needs help at all to resume after sleep. (Let's see how to make this better after. Anyway, the whole code of the lighting service is a bit shitty, as it was the first one to be written)