🔥 Start getting rid of GRPC in the overlay program.
Much to be done still, but this introduces a working skeleton for the low-level pipe-based client and servers that are going to replace GRPC.
Overlay is being done first as it is obviously the easiest one. But more importantly, outside of getting rid of a lot of bulk, it should allow for full AOT compilation of that component.