🐛 Do not limit the number of concurrent instances for now.
I was maybe too optimistic when thinking about this initially.
It would be easy to have a different loop for the case where we only want a single client. (Which may be the only one that matters; would only be annoying for debugging if anything)
However, handling the general case of a limitation is maybe not that easy. It could probably be done through a semaphore, but that is not an urgent change.