⚙️ Add delay and more retries when retrieving monitor capabilities through the Windows API.
This should help mitigate issues that sometimes arise when a monitor has just been connected. (Or if the input of said monitor has just been switched to the computer)
The "fix" is not ideal, but we need to either expose the retry logic outside, or return a more actionable exception so that callers can decide themselves what to do.
A possible imrovement would possibly be to skip the call to GetCapabilitiesStringLength and directly provide a relatively large buffer (e.g. 4k or 8k) with hope that the capabilitis string fits inside. (Should be most of the time)
Otherwise, I fear we are actually querying the capabilities at least two times due to the way the API is designed.