✨ Rely on System.Text.Json source generation to optimize the Elgato Lights driver.
Given the simplicity of the protocol, I might consider going full custom for the JSON encoding and decoding of regular requests. Especially as the current source generation approach does not cover deserialization.
However, there will still be the problem of the periodic HTTP GET requests themselves, which seems to be far from free. (Certainly acceptable, but I wish I'd be able to make them close to alloc free)