✨ Try to make the sensor watching a bit lighter.
From what I observed, on a compter with a lot of sensors, watching can cause quite "a lot" of garbage to be generated every second. (Up to some kind of plateau)
As a first step, reducing the over-reliance on IAsyncEnumerable should be helpful at least on the performance side (less jumping around interface implementations).