✨ Move out the rest of the sensor API to the new protocol.
This also comes with a slight memory management improvement. A VariantNumber struct that can store most numerical data types is introduced so that values don't require to be boxed or converted to double at many places.
Values are still converted to double in the UI for convenience (also it doesn't matter except for very large integers), but they are transmitted as-is on the wire.