♻️ Improve the architecture of the new IPC code in the UI.
The full asynchronous model that was used for the Helper would not work well for the main UI as we need to have strict ordrering of connect/disocnnect operations.
This changes the code to a more directive model with less IoC but it should make most of the code simple, not having to implement many independent "watchers" everywhere.
This will unlock migrating other messages to the new protocol.