✨ Add a prototype source-generator-based effect serializer.
For now this is not used and it currently relies on a default implementation in the ISerializer interface (which works for all effects because all are blittable)
The goal is to entirely replace the reflection-based effect serializer with this new logic by also pre-generating metadata.
Metadata format will need to handle fields/properties of non primitive type (blittable too) so we'll so how it gets done. (Maybe the known data types will be hardcoded at first ?)
Having a proper definition will also allow the UI to entirely work on binary data rather than the clunky semi-high-level objects that are currently passed through protobuf.