🐛 Fix bugs introduced in the last commits.
With this, things seem to mostly work as they did previously.
More tests to be done.
Optionally implement a migration process if necessary, but at this stage, losing defiend effects is probably ok. (Old ones won't be erased anyway, so that it is still possible to read them manually and redo them in the new format)