🐛 Lighting service: Do push new devices 🙂
Hard to notice if the service already has a configuration, but the UI may be missing some information.
There is still a problem for when a device gets new effects (i.e. after driver improvement and the device comes online), but IIRC this one is purely on UI side (not implemented).