✨ Migrate existing multi-color effects to use color arrays, fixed or variable.
Sadly, this will break currently serialized values for effects with a variable number of colors, but it makes everything cleaner in the end, including the UI.
Also rename pulse to breathing for razer, as the effect is clearly more breathing than pulse (also that's what razer calls it, I think).
My bad for confusing the two effects (altough there is an obvious similarity).