✨ Start adding conversion operators to some effect types.
The end goal is to allow automatic conversion at assignment time.
It is *one* of the way to avoid multiplying the list of supported effects on every lighting zone type for each set of supported parameters.
Another way would be to be able to separate the effect from its parameters, which would be nice but it also comes with its own sets of problems.