🐛 Cleanup cancellations from the wwaiters in AsyncLock.
Not sure if that's the only reason, but when debugging locally, I noticed quite a few CancellationTokenSource.CallbackNode alive.
It seems at least a few of them were traceable to AsyncLock, and indeed, the method to clear registrations is not always called.
So I assume that over time, when starting and stopping the UI several times (probably over 50 times in my case) some registrations could be queued on some forever-alive cancellation tokens. (Only in case of lock contention though)