♻️ Migrate the MonitorService to IChangeSource.
This should reduce a bit the garbage kept alive in the process.
It seems that some instances of ChannelWriter were being kept alive after UI disconnection somehow. The newer mechanism is more robust is used properly as it is used by most services, and will avoid these kind of issues by not relying on async enumerators.