✨ Read more SMBIOS structures.
There are still a lot of structures that are not parsed by the library. I added the most basic missing ones for now.

Being able to retrieve info on the motherboard might be useful to implement proper motherboard drivers.
To be seen how to proceed, but one of the idea is to use the basebaord info from SMBIOS to choose one and exactly one driver to instantiate as the motherboard driver. This might not be perfect as motherboards can still have many attached components, some of which we may still want to expose independently? Frankly, I don't know yet.