♻️ Try to improve monitor stuff.
Somehow, some null monitor messages are received on the connection. Trying to understand why and how.
It doesn't reproduce between a debug instance of the service and a debug anstance of the UI. However it does occurr from a release instance of the server to *some* versions of the UI?