♻️ Improve pipe connections. Memory usage + not logging PipeClosed.
Should save a few dozen bytes of memory per ui connection, plus and minus the differences in resulting code, which should be more savings on average. (At least 108 bytes, maybe up to 0.5kB total if we account for the removed interfaces and JIT work)
This is not a lot compared to the global consumption, but it is a very easy change, so it should be done.