✨ Try to make the polling scheduler less reliant on allocations.
One of the main goal is to reduce the overhead of TaskCanceledException popping when the UI is closed.
This sets up the ground work for cancelling those operations without relying on exceptions as much.
It should also avoid one TaskCompletionSource allocation per tick. The downside being that slow sensors may seemingly miss their ticks more easily. (Should have also happened with the previous implementation, so I will investigate a bit)