🐛 Try to fix the line chart component sometimes not correctly registering a scale change.
Symptom observed: when the whole curve (60s of data points) is displayed, then the missing scales appear. (Assumption is that there are some values that cause an invalid computation of min and max. But why not always, I have no idea.)

TBH: Should probably backport much of the code from the power control curve editor over there (which would allow for scales) but 💤