♻️ Improve handling of Elgato Led Strip Pro.
* Revert update packets to using integers for H/S/B.
* Proper conversions for expanded hue/saturation units to integer percentage.
* Split all "informative" lighting properties into the persistence feature.
* Add a feature for reacting to dynamic brightness changes.
* Add custom names to effects set on the Light Strip Pro.
* Fixed brightness updates. (Apparently it wouldn't work without setting the "numberOfLights" property?)

All of this is probably still not perfect and there are likely some possible race conditions in the code.
This is unlikely to cause any big problem, so this will be addressed later. In the end, we will probably need to enable deferred lighting changes in the driver, which will make it more complex, but will also allow to solve this problem more easily.