♻️ Move lighting effects into a dedicated assembly.
Was initially trying to do this so that the EffectSerializer infrastructure didn't interfere with monitor definition compiler in case of an error during dev, but it won't solve this problem.
It will still be better in the long run, as the number of effects keep increasing.
Long term plan is to have most effects as plugins so that anyone can provide ARGB effects by adding in a new library.
Of course, core effects will always be present, as they are a dependency of many lighting devices, but others could be made optional.